requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==4.2
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
psycopg2==2.9.10
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
